rules/rule-service/pyproject.toml
[tool.poetry]
name = "rule-service"
version = "0.1.0"
description = ""
authors = ["David Papp <pigri@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
transformers = "^4.42.4"
torch = "^2.3.1"
rule-engine = "^4.5.0"
presidio-analyzer = "^2.2.355"
presidio-anonymizer = "^2.2.355"
requests = "^2.32.3"
uvicorn = "^0.30.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
